@@ -1,215 +1,417 @@
 {
+  "name": "docs.github.com",
+  "private": true,
+  "repository": "https://github.com/github/docs",
+  "license": "(MIT AND CC-BY-4.0)",
   "author": {
-    "email": "opensource+docs@github.com",
     "name": "GitHub",
+    "email": "opensource+docs@github.com",
     "url": "https://github.com/github/docs"
+  },
+  "type": "module",
+  "imports": {
+    "#src/*": {
+      "default": "./src/*"
+    }
+  },
+  "exports": "./src/frame/server.ts",
+  "scripts": {
+    "all-documents": "tsx src/content-render/scripts/all-documents/cli.ts",
+    "analyze-text": "tsx src/search/scripts/analyze-text.ts",
+    "analyze-comment": "tsx src/events/scripts/analyze-comment-cli.ts",
+    "archive-version": "tsx --max-old-space-size=16384 src/ghes-releases/scripts/archive-version.ts",
+    "build": "next build",
+    "check-content-type": "tsx src/workflows/check-content-type.ts",
+    "check-github-github-links": "tsx src/links/scripts/check-github-github-links.ts",
+    "cmp-files": "tsx src/workflows/cmp-files.ts",
+    "content-changes-table-comment": "tsx src/workflows/content-changes-table-comment.ts",
+    "copy-fixture-data": "tsx src/tests/scripts/copy-fixture-data.js",
+    "count-translation-corruptions": "tsx src/languages/scripts/count-translation-corruptions.ts",
+    "create-enterprise-issue": "tsx src/ghes-releases/scripts/create-enterprise-issue.js",
+    "debug": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon --inspect src/frame/server.ts",
+    "delete-orphan-translation-files": "tsx src/workflows/delete-orphan-translation-files.ts",
+    "deleted-assets-pr-comment": "tsx src/assets/scripts/deleted-assets-pr-comment.ts",
+    "deleted-features-pr-comment": "tsx src/data-directory/scripts/deleted-features-pr-comment.ts",
+    "dev": "cross-env npm start",
+    "enable-automerge": "tsx src/workflows/enable-automerge.ts",
+    "find-orphaned-assets": "tsx src/assets/scripts/find-orphaned-assets.ts",
+    "find-orphaned-features": "tsx src/data-directory/scripts/find-orphaned-features/index.ts",
+    "find-past-built-pr": "tsx src/workflows/find-past-built-pr.ts",
+    "find-unused-variables": "tsx src/content-linter/scripts/find-unsed-variables.ts",
+    "fixture-dev": "cross-env ROOT=src/fixtures/fixtures npm start",
+    "fixture-test": "cross-env ROOT=src/fixtures/fixtures npm test -- src/fixtures/tests",
+    "fr-add-docs-reviewers-requests": "tsx src/workflows/fr-add-docs-reviewers-requests.ts",
+    "general-search-scrape": "tsx src/search/scripts/scrape/scrape-cli.ts",
+    "general-search-scrape-server": "cross-env NODE_ENV=production PORT=4002 MINIMAL_RENDER=true CHANGELOG_DISABLED=true tsx src/frame/server.ts",
+    "ghes-release-scrape-with-server": "cross-env GHES_RELEASE=1 start-server-and-test general-search-scrape-server 4002 general-search-scrape",
+    "general-search-scrape-with-server": "cross-env NODE_OPTIONS='--max_old_space_size=8192' start-server-and-test general-search-scrape-server 4002 general-search-scrape",
+    "index": "tsx src/search/scripts/index/index-cli autocomplete docs-internal-data",
+    "index-ai-search-autocomplete": "tsx src/search/scripts/index/index-cli ai-search-autocomplete",
+    "index-general-autocomplete": "tsx src/search/scripts/index/index-cli general-autocomplete",
+    "index-general-search": "tsx src/search/scripts/index/index-cli general-search",
+    "index-test-fixtures": "./src/search/scripts/index-test-fixtures.sh",
+    "labeler": "tsx .github/actions/labeler/labeler.ts",
+    "lint": "eslint '**/*.{js,mjs,ts,tsx}'",
+    "lint-content": "tsx src/content-linter/scripts/lint-content.js",
+    "lint-translation": "vitest src/content-linter/tests/lint-files.js",
+    "liquid-markdown-tables": "tsx src/tools/scripts/liquid-markdown-tables/index.ts",
+    "generate-code-scanning-query-list": "tsx src/code-scanning/scripts/generate-code-scanning-query-list.ts",
+    "generate-content-linter-docs": "tsx src/content-linter/scripts/generate-docs.ts",
+    "move-content": "tsx src/content-render/scripts/move-content.js",
+    "openapi-docs": "tsx src/rest/docs.js",
+    "playwright-test": "playwright test --config src/fixtures/playwright.config.ts --project=\"Google Chrome\"",
+    "post-lints": "tsx src/content-linter/scripts/post-lints.js",
+    "postinstall": "cp package-lock.json .installed.package-lock.json && echo \"Updated .installed.package-lock.json\" # see husky/post-checkout and husky/post-merge",
+    "precompute-pageinfo": "tsx src/pageinfo/scripts/precompute-pageinfo.ts",
+    "prepare": "husky src/workflows/husky",
+    "prettier": "prettier -w \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
+    "prettier-check": "prettier -c \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
+    "prevent-pushes-to-main": "tsx src/workflows/prevent-pushes-to-main.ts",
+    "purge-fastly-edge-cache": "tsx src/workflows/purge-fastly-edge-cache.ts",
+    "purge-fastly-edge-cache-per-language": "tsx src/languages/scripts/purge-fastly-edge-cache-per-language.js",
+    "purge-old-workflow-runs": "tsx src/workflows/purge-old-workflow-runs.js",
+    "ready-for-docs-review": "tsx src/workflows/ready-for-docs-review.ts",
+    "release-banner": "tsx src/ghes-releases/scripts/release-banner.js",
+    "reusables": "tsx src/content-render/scripts/reusables-cli.ts",
+    "remove-version-markup": "tsx src/ghes-releases/scripts/remove-version-markup.js",
+    "rendered-content-link-checker": "tsx src/links/scripts/rendered-content-link-checker.ts",
+    "rendered-content-link-checker-cli": "tsx src/links/scripts/rendered-content-link-checker-cli.ts",
+    "rest-dev": "tsx src/rest/scripts/update-files.js",
+    "show-action-deps": "echo 'Action Dependencies:' && rg '^[\\s|-]*(uses:.*)$' .github -I -N --no-heading -r '$1$2' | sort | uniq | cut -c 7-",
+    "start": "cross-env NODE_ENV=development ENABLED_LANGUAGES=en nodemon src/frame/server.ts",
+    "start-all-languages": "cross-env NODE_ENV=development tsx src/frame/server.ts",
+    "start-for-playwright": "cross-env ROOT=src/fixtures/fixtures TRANSLATIONS_FIXTURE_ROOT=src/fixtures/fixtures/translations ENABLED_LANGUAGES=en,ja NODE_ENV=test tsx src/frame/server.ts",
+    "symlink-from-local-repo": "tsx src/early-access/scripts/symlink-from-local-repo.js",
+    "sync-audit-log": "tsx src/audit-logs/scripts/sync.ts",
+    "sync-graphql": "tsx src/graphql/scripts/sync.js",
+    "sync-rest": "tsx src/rest/scripts/update-files.ts",
+    "sync-secret-scanning": "tsx src/secret-scanning/scripts/sync.ts",
+    "sync-webhooks": "npx tsx src/rest/scripts/update-files.ts -o webhooks",
+    "test": "vitest",
+    "test-local-dev": "tsx src/workflows/test-local-dev.ts",
+    "test-moved-content": "tsx src/content-render/scripts/test-moved-content.ts",
+    "tsc": "tsc --noEmit",
+    "unallowed-contributions": "tsx src/workflows/unallowed-contributions.ts",
+    "update-data-and-image-paths": "tsx src/early-access/scripts/update-data-and-image-paths.js",
+    "update-enterprise-dates": "tsx src/ghes-releases/scripts/update-enterprise-dates.js",
+    "update-internal-links": "tsx src/links/scripts/update-internal-links.ts",
+    "validate-asset-images": "tsx src/assets/scripts/validate-asset-images.ts",
+    "validate-github-github-docs-urls": "tsx src/links/scripts/validate-github-github-docs-urls/index.ts",
+    "warmup-remotejson": "tsx src/archives/scripts/warmup-remotejson.ts",
+    "what-docs-early-access-branch": "tsx src/early-access/scripts/what-docs-early-access-branch.js"
+  },
+  "lint-staged": {
+    "*.{js,mjs,ts,tsx}": "eslint --cache --fix",
+    "*.{js,mjs,scss,ts,tsx,yml,yaml}": "prettier --write",
+    "{content,data}/**/*.md": "npm run lint-content -- --precommit --paths"
+  },
+  "nodemonConfig": {
+    "ext": "ts,js,json,yml,md,html,scss",
+    "exec": "tsx",
+    "ignore": [
+      "assets",
+      "script",
+      "tests",
+      "content",
+      "translations",
+      "data/reusables",
+      "data/variables",
+      "data/glossaries",
+      "data/product-examples",
+      "data/learning-tracks",
+      "rest-api-description",
+      "semmle-code"
+    ]
+  },
+  "prettier": {
+    "overrides": [
+      {
+        "files": [
+          "**/*.{yml,yaml}"
+        ],
+        "options": {
+          "singleQuote": true
+        }
+      },
+      {
+        "files": [
+          "**/*.{ts,tsx,js,mjs}"
+        ],
+        "options": {
+          "printWidth": 100,
+          "semi": false,
+          "singleQuote": true
+        }
+      }
+    ]
+  },
+  "eslintConfig": {
+    "env": {
+      "browser": true,
+      "commonjs": true,
+      "es2020": true,
+      "node": true
+    },
+    "parserOptions": {
+      "ecmaVersion": 2022,
+      "requireConfigFile": "false",
+      "sourceType": "module"
+    },
+    "settings": {
+      "import/resolver": {
+        "typescript": true,
+        "node": true
+      }
+    },
+    "extends": [
+      "eslint:recommended",
+      "plugin:github/recommended",
+      "plugin:import/errors",
+      "prettier"
+    ],
+    "rules": {
+      "import/no-extraneous-dependencies": [
+        "error",
+        {
+          "packageDir": "."
+        }
+      ],
+      "import/extensions": "off",
+      "no-console": "off",
+      "github/array-foreach": "off",
+      "camelcase": "off",
+      "i18n-text/no-en": "off",
+      "no-shadow": "off",
+      "prefer-template": "off",
+      "filenames/match-regex": "off",
+      "no-constant-condition": "off",
+      "no-unused-vars": "off",
+      "github/no-then": "off",
+      "import/no-named-as-default-member": "off",
+      "one-var": "off",
+      "import/no-namespace": "off",
+      "import/no-anonymous-default-export": "off",
+      "object-shorthand": "off",
+      "eslint-comments/no-use": "off",
+      "no-empty": "off",
+      "prefer-const": "off",
+      "import/no-named-as-default": "off",
+      "eslint-comments/no-unused-disable": "off",
+      "no-useless-concat": "off",
+      "func-style": "off",
+      "eslint-comments/no-unlimited-disable": "off"
+    },
+    "overrides": [
+      {
+        "files": [
+          "**/*.tsx",
+          "**/*.ts"
+        ],
+        "parser": "@typescript-eslint/parser",
+        "plugins": [
+          "@typescript-eslint",
+          "primer-react",
+          "jsx-a11y"
+        ],
+        "extends": [
+          "plugin:primer-react/recommended",
+          "plugin:jsx-a11y/recommended"
+        ],
+        "rules": {
+          "camelcase": "off",
+          "no-undef": "off",
+          "no-unused-vars": "off",
+          "no-use-before-define": "off",
+          "@typescript-eslint/no-unused-vars": "error",
+          "jsx-a11y/no-onchange": "off"
+        }
+      }
+    ],
+    "ignorePatterns": [
+      "tmp/*",
+      "!/.*",
+      "/.next/",
+      "src/bookmarklets/*",
+      "rest-api-description/",
+      "docs-internal-data/",
+      "src/code-scanning/scripts/generate-code-scanning-query-list.ts"
+    ]
+  },
+  "dependencies": {
+    "@elastic/elasticsearch": "8.13.1",
+    "@github/failbot": "0.8.3",
+    "@horizon-rs/language-guesser": "0.1.1",
+    "@octokit/plugin-retry": "7.1.2",
+    "@octokit/request-error": "6.1.1",
+    "@primer/behaviors": "^1.7.0",
+    "@primer/css": "^21.3.1",
+    "@primer/live-region-element": "^0.7.0",
+    "@primer/octicons": "^19.14.0",
+    "@primer/octicons-react": "^19.14.0",
+    "@primer/react": "36.27.0",
+    "accept-language-parser": "^1.5.0",
+    "ajv": "^8.17.1",
+    "ajv-errors": "^3.0.0",
+    "ajv-formats": "^3.0.1",
+    "bottleneck": "2.19.5",
+    "boxen": "8.0.1",
+    "cheerio": "^1.0.0-rc.11",
+    "cheerio-to-text": "0.2.4",
+    "classnames": "^2.5.1",
+    "connect-timeout": "1.9.0",
+    "cookie-parser": "^1.4.7",
+    "cuss": "2.2.0",
+    "dayjs": "^1.11.3",
+    "dotenv": "^16.4.5",
+    "escape-string-regexp": "5.0.0",
+    "express": "4.21.2",
+    "express-rate-limit": "7.4.0",
+    "fastest-levenshtein": "1.0.16",
+    "file-type": "20.0.0",
+    "flat": "^6.0.1",
+    "github-slugger": "^2.0.0",
+    "glob": "11.0.0",
+    "got": "^14.4.2",
+    "gray-matter": "^4.0.3",
+    "hast-util-from-parse5": "^8.0.1",
+    "hast-util-to-string": "^3.0.1",
+    "hastscript": "^9.0.0",
+    "helmet": "^8.0.0",
+    "highlight.js": "11.9.0",
+    "highlightjs-curl": "^1.3.0",
+    "hot-shots": "^10.0.0",
+    "html-entities": "^2.5.2",
+    "http-proxy-middleware": "3.0.3",
+    "imurmurhash": "^0.1.4",
+    "is-svg": "5.0.0",
+    "javascript-stringify": "^2.1.0",
+    "js-cookie": "^3.0.1",
+    "js-yaml": "^4.1.0",
+    "liquidjs": "^10.16.1",
+    "lodash": "^4.17.21",
+    "lodash-es": "^4.17.21",
+    "lowdb": "7.0.1",
+    "lowlight": "3.1.0",
+    "markdownlint-rule-helpers": "^0.25.0",
+    "mdast-util-from-markdown": "^2.0.0",
+    "mdast-util-to-hast": "^13.2.0",
+    "mdast-util-to-markdown": "2.1.0",
+    "mdast-util-to-string": "^4.0.0",
+    "morgan": "^1.10.0",
+    "next": "^15.1.5",
+    "ora": "^8.0.1",
+    "parse5": "7.1.2",
+    "quick-lru": "7.0.0",
+    "react": "18.3.1",
+    "react-dom": "18.3.1",
+    "react-markdown": "^9.0.1",
+    "rehype-highlight": "^7.0.0",
+    "rehype-raw": "^7.0.0",
+    "rehype-slug": "^6.0.0",
+    "rehype-stringify": "^10.0.0",
+    "remark-gemoji-to-emoji": "^1.1.0",
+    "remark-gfm": "^4.0.0",
+    "remark-parse": "^11.0.0",
+    "remark-rehype": "^11.1.0",
+    "remark-remove-comments": "^1.0.1",
+    "remark-stringify": "^11.0.0",
+    "rss-parser": "^3.13.0",
+    "scroll-anchoring": "^0.1.0",
+    "semver": "^7.6.2",
+    "sharp": "0.33.1",
+    "slash": "^5.1.0",
+    "strip-ansi": "7.1.0",
+    "styled-components": "^5.3.5",
+    "swr": "^2.2.4",
+    "tcp-port-used": "1.0.2",
+    "tsx": "4.15.1",
+    "unified": "^11.0.3",
+    "unist-util-visit": "^5.0.0",
+    "url-template": "^3.1.1",
+    "walk-sync": "^3.0.0"
+  },
+  "devDependencies": {
+    "@actions/core": "^1.10.0",
+    "@actions/github": "^6.0.0",
+    "@axe-core/playwright": "^4.10.0",
+    "@github/markdownlint-github": "^0.6.2",
+    "@graphql-inspector/core": "^6.1.0",
+    "@graphql-tools/load": "^8.0.0",
+    "@octokit/rest": "21.0.2",
+    "@playwright/test": "^1.50",
+    "@types/accept-language-parser": "1.5.6",
+    "@types/cheerio": "^0.22.35",
+    "@types/connect-timeout": "0.0.39",
+    "@types/cookie": "0.6.0",
+    "@types/cookie-parser": "1.4.7",
+    "@types/elasticsearch": "^5.0.43",
+    "@types/event-to-promise": "^0.7.5",
+    "@types/express": "4.17.21",
+    "@types/imurmurhash": "^0.1.4",
+    "@types/js-cookie": "^3.0.6",
+    "@types/js-yaml": "^4.0.9",
+    "@types/lodash": "^4.17.0",
+    "@types/lodash-es": "4.17.12",
+    "@types/mdast": "^4.0.4",
+    "@types/morgan": "1.9.9",
+    "@types/react": "18.3.3",
+    "@types/react-dom": "^18.3.0",
+    "@types/semver": "^7.5.8",
+    "@types/styled-components": "^5.1.34",
+    "@types/tcp-port-used": "1.0.4",
+    "@types/website-scraper": "^1.2.10",
+    "@typescript-eslint/eslint-plugin": "^8.7.0",
+    "@typescript-eslint/parser": "^8.7.0",
+    "chalk": "^5.0.1",
+    "change-case": "^5.4.4",
+    "commander": "^13.0.0",
+    "cross-env": "^7.0.3",
+    "csp-parse": "0.0.2",
+    "csv-parse": "5.5.6",
+    "eslint": "^8.57.1",
+    "eslint-config-prettier": "^10.0.1",
+    "eslint-import-resolver-typescript": "^3.6.3",
+    "eslint-plugin-github": "^5.0.2",
+    "eslint-plugin-import": "^2.30.0",
+    "eslint-plugin-jsx-a11y": "^6.10.0",
+    "eslint-plugin-primer-react": "^6.1.2",
+    "event-to-promise": "^0.8.0",
+    "graphql": "^16.9.0",
+    "http-status-code": "^2.1.0",
+    "husky": "^9.1.4",
+    "json-schema-merge-allof": "^0.8.1",
+    "kill-port": "2.0.1",
+    "lint-staged": "^15.2.9",
+    "markdownlint": "^0.34.0",
+    "markdownlint-rule-search-replace": "^1.2.0",
+    "mdast-util-gfm-table": "^2.0.0",
+    "micromark-extension-gfm-table": "^2.0.0",
+    "mkdirp": "^3.0.0",
+    "mockdate": "^3.0.5",
+    "nock": "^14.0.0",
+    "nodemon": "3.1.3",
+    "npm-merge-driver-install": "^3.0.0",
+    "nth-check": "2.1.1",
+    "prettier": "^3.3.2",
+    "rimraf": "^6.0.0",
+    "robots-parser": "^3.0.0",
+    "sass": "^1.77.1",
+    "start-server-and-test": "^2.0.3",
+    "typescript": "^5.5.2",
+    "unist-util-remove": "^4.0.0",
+    "unist-util-visit-parents": "6.0.1",
+    "vitest": "3.0.5",
+    "website-scraper": "^5.3.1"
+  },
+  "overrides": {},
+  "optionalDependencies": {
+    "esm": "^3.2.25"
+  },
+  "engines": {
+    "node": "^20 || ^22"
   },
   "cacheDirectories": [
     "node_modules",
     ".next/cache"
-  ],
-  "dependencies": {
-    "@elastic/elasticsearch": "8.2.1",
-    "@github/failbot": "0.8.0",
-    "@primer/css": "^20.2.4",
-    "@primer/octicons": "17.3.0",
-    "@primer/octicons-react": "17.3.0",
-    "@primer/react": "^35.2.2",
-    "accept-language-parser": "^1.5.0",
-    "ajv": "^8.11.0",
-    "ajv-formats": "^2.1.1",
-    "cheerio": "^1.0.0-rc.11",
-    "classnames": "^2.3.1",
-    "connect-datadog": "0.0.9",
-    "cookie-parser": "^1.4.6",
-    "cors": "^2.8.5",
-    "csurf": "^1.11.0",
-    "dayjs": "^1.11.3",
-    "dotenv": "^16.0.1",
-    "express": "^4.18.1",
-    "express-timeout-handler": "^2.2.2",
-    "flat": "^5.0.2",
-    "github-slugger": "^1.4.0",
-    "got": "^12.1.0",
-    "gray-matter": "^4.0.3",
-    "hast-util-from-parse5": "^7.1.0",
-    "hast-util-parse-selector": "^3.1.0",
-    "hast-util-select": "^5.0.2",
-    "hast-util-to-string": "^2.0.0",
-    "hastscript": "^7.0.2",
-    "helmet": "^5.1.0",
-    "highlight.js": "11.5.1",
-    "highlightjs-curl": "^1.3.0",
-    "highlightjs-graphql": "^1.0.2",
-    "hot-shots": "^9.0.0",
-    "html-entities": "^2.3.3",
-    "imurmurhash": "^0.1.4",
-    "javascript-stringify": "^2.1.0",
-    "js-cookie": "^3.0.1",
-    "js-yaml": "^4.1.0",
-    "liquidjs": "9.22.1",
-    "lodash": "^4.17.21",
-    "lodash-es": "^4.17.21",
-    "lunr": "^2.3.9",
-    "lunr-languages": "^1.9.0",
-    "mdast-util-from-markdown": "^1.2.0",
-    "mdast-util-to-string": "^3.1.0",
-    "morgan": "^1.10.0",
-    "msgpack5rpc": "^1.1.0",
-    "next": "12.1.6",
-    "overload-protection": "^1.2.3",
-    "parse5": "7.0.0",
-    "port-used": "^2.0.8",
-    "quick-lru": "6.1.1",
-    "react": "^17.0.2",
-    "react-dom": "^17.0.2",
-    "react-markdown": "^8.0.3",
-    "react-syntax-highlighter": "^15.5.0",
-    "rehype-autolink-headings": "^6.1.1",
-    "rehype-highlight": "^5.0.2",
-    "rehype-raw": "^6.1.1",
-    "rehype-slug": "^5.0.1",
-    "rehype-stringify": "^9.0.3",
-    "remark-code-extra": "^1.0.1",
-    "remark-gemoji-to-emoji": "^1.1.0",
-    "remark-gfm": "^3.0.1",
-    "remark-parse": "^10.0.1",
-    "remark-parse-no-trim": "^8.0.4",
-    "remark-rehype": "^10.1.0",
-    "revalidator": "^0.3.1",
-    "rss-parser": "^3.12.0",
-    "scroll-anchoring": "^0.1.0",
-    "semver": "^7.3.7",
-    "slash": "^4.0.0",
-    "strip-html-comments": "^1.0.0",
-    "styled-components": "^5.3.5",
-    "swr": "1.3.0",
-    "ts-dedent": "^2.2.0",
-    "unified": "^10.1.2",
-    "unist-util-visit": "^4.1.0",
-    "url-template": "^3.0.0",
-    "uuid": "^8.3.2",
-    "walk-sync": "^3.0.0"
-  },
-  "devDependencies": {
-    "@actions/core": "^1.9.0",
-    "@actions/github": "^5.0.3",
-    "@alex_neo/jest-expect-message": "^1.0.5",
-    "@babel/core": "^7.18.5",
-    "@babel/eslint-parser": "^7.18.2",
-    "@babel/plugin-syntax-top-level-await": "^7.14.5",
-    "@babel/plugin-transform-runtime": "^7.18.5",
-    "@babel/preset-env": "^7.18.2",
-    "@graphql-inspector/core": "^3.1.2",
-    "@graphql-tools/load": "^7.5.14",
-    "@jest/globals": "^28.1.1",
-    "@octokit/graphql": "4.8.0",
-    "@octokit/rest": "^18.12.0",
-    "@types/github-slugger": "^1.3.0",
-    "@types/imurmurhash": "^0.1.1",
-    "@types/js-cookie": "^3.0.2",
-    "@types/lodash": "^4.14.182",
-    "@types/react": "17.0.38",
-    "@types/react-dom": "^18.0.5",
-    "@types/react-syntax-highlighter": "^15.5.2",
-    "@types/uuid": "^8.3.4",
-    "@typescript-eslint/eslint-plugin": "5.30.7",
-    "@typescript-eslint/parser": "5.30.7",
-    "babel-loader": "^8.2.5",
-    "babel-plugin-styled-components": "^2.0.7",
-    "babel-preset-env": "^1.7.0",
-    "chalk": "^5.0.1",
-    "change-case": "^4.1.2",
-    "commander": "^9.3.0",
-    "cross-env": "^7.0.3",
-    "csp-parse": "0.0.2",
-    "dedent": "^0.7.0",
-    "domwaiter": "^1.3.0",
-    "eslint": "8.20.0",
-    "eslint-config-prettier": "^8.5.0",
-    "eslint-config-standard": "^17.0.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsx-a11y": "^6.5.1",
-    "eslint-plugin-node": "^11.1.0",
-    "eslint-plugin-promise": "^6.0.0",
-    "event-to-promise": "^0.8.0",
-    "git-diff": "^2.0.6",
-    "glob": "^8.0.3",
-    "graphql": "^16.5.0",
-    "http-status-code": "^2.1.0",
-    "husky": "^8.0.1",
-    "japanese-characters": "^1.1.0",
-    "jest": "28.1.3",
-    "jest-environment-puppeteer": "^5.0.4",
-    "jest-fail-on-console": "^2.4.2",
-    "jest-github-actions-reporter": "^1.0.3",
-    "jest-slow-test-reporter": "^1.0.0",
-    "kill-port": "2.0.0",
-    "linkinator": "^3.1.0",
-    "lint-staged": "^13.0.1",
-    "make-promises-safe": "^5.1.0",
-    "minimatch": "^5.1.0",
-    "mkdirp": "^1.0.4",
-    "mockdate": "^3.0.5",
-    "nock": "^13.2.7",
-    "nodemon": "2.0.19",
-    "npm-merge-driver-install": "^3.0.0",
-    "nth-check": "2.1.1",
-    "postcss": "^8.4.14",
-    "prettier": "^2.7.0",
-    "replace": "^1.2.1",
-    "rimraf": "^3.0.2",
-    "robots-parser": "^3.0.0",
-    "sass": "^1.52.3",
-    "start-server-and-test": "^1.14.0",
-    "strip-ansi": "^7.0.1",
-    "tough-cookie": "4.0.0",
-    "typescript": "^4.7.3"
-  },
-  "engines": {
-    "node": ">=16.x"
-  },
-  "exports": "./server.mjs",
-  "license": "(MIT AND CC-BY-4.0)",
-  "name": "docs.github.com",
-  "optionalDependencies": {
-    "bottleneck": "^2.19.5",
-    "esm": "^3.2.25",
-    "image-size": "^1.0.1",
-    "jest-puppeteer": "^5.0.4",
-    "puppeteer": "^9.1.1",
-    "website-scraper": "^5.0.0"
-  },
-  "private": true,
-  "repository": "https://github.com/github/docs",
-  "scripts": {
-    "browser-test": "cross-env BROWSER=1 NODE_OPTIONS=--experimental-vm-modules jest tests/browser/browser.js",
-    "build": "next build",
-    "debug": "cross-env NODE_ENV=development ENABLED_LANGUAGES='en,ja' nodemon --inspect server.mjs",
-    "dev": "npm start",
-    "lint": "eslint '**/*.{js,mjs,ts,tsx}'",
-    "lint-translation": "cross-env NODE_OPTIONS=--experimental-vm-modules TEST_TRANSLATION=true jest tests/linting/lint-files.js",
-    "prepare": "husky install",
-    "prettier": "prettier -w \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
-    "prettier-check": "prettier -c \"**/*.{ts,tsx,js,mjs,scss,yml,yaml}\"",
-    "prevent-pushes-to-main": "node script/prevent-pushes-to-main.js",
-    "rest-dev": "script/rest/update-files.js && npm run dev",
-    "show-action-deps": "echo 'Action Dependencies:' && rg '^[\\s|-]*(uses:.*)$' .github -I -N --no-heading -r '$1$2' | sort | uniq | cut -c 7-",
-    "start": "cross-env NODE_ENV=development ENABLED_LANGUAGES='en,ja' nodemon server.mjs",
-    "start-all-languages": "cross-env NODE_ENV=development nodemon server.mjs",
-    "sync-search": "cross-env NODE_OPTIONS='--max_old_space_size=8192' start-server-and-test sync-search-server 4002 sync-search-indices",
-    "sync-search-ghes-release": "cross-env GHES_RELEASE=1 start-server-and-test sync-search-server 4002 sync-search-indices",
-    "sync-search-indices": "script/search/sync-search-indices.js",
-    "sync-search-server": "cross-env NODE_ENV=production PORT=4002 CHANGELOG_DISABLED=true DISABLE_RENDERING_CACHE=true DISABLE_OVERLOAD_PROTECTION=true node server.mjs",
-    "translation-check": "start-server-and-test translation-check-server 4002 translation-check-test",
-    "translation-check-server": "cross-env NODE_ENV=test PORT=4002 node server.mjs",
-    "translation-check-test": "script/i18n/test-html-pages.js",
-    "test": "cross-env NODE_OPTIONS='--max_old_space_size=4096 --experimental-vm-modules' jest --logHeapUsage",
-    "test-watch": "cross-env NODE_OPTIONS=--experimental-vm-modules jest --watch --notify --notifyMode=change --coverage",
-    "tsc": "tsc --noEmit"
-  },
-  "lint-staged": {
-    "*.{js,mjs,ts,tsx}": "eslint --cache --fix",
-    "*.{js,mjs,scss,ts,tsx,yml,yaml}": "prettier --write"
-  }
+  ]
 }
