@@ -1,14 +1,15 @@
-// import { productIds } from './lib/all-products.js'
-// import languages from './lib/languages.js'
+import fs from 'fs'
+import path from 'path'
 
-const fs = require('fs')
-const frontmatter = require('gray-matter')
-const path = require('path')
-const homepage = path.posix.join(process.cwd(), 'content/index.md')
+import frontmatter from 'gray-matter'
+import { languageKeys } from '#src/languages/lib/languages.js'
+import { ROOT } from '#src/frame/lib/constants.js'
+
+const homepage = path.posix.join(ROOT, 'content/index.md')
 const { data } = frontmatter(fs.readFileSync(homepage, 'utf8'))
 const productIds = data.children
 
-module.exports = {
+export default {
   // speed up production `next build` by ignoring typechecking during that step of build.
   // type-checking still occurs in the Dockerfile build
   typescript: {
@@ -18,8 +19,7 @@ module.exports = {
     ignoreDuringBuilds: true,
   },
   i18n: {
-    // locales: Object.values(languages).map(({ code }) => code),
-    locales: ['en', 'cn', 'ja', 'es', 'pt'],
+    locales: languageKeys,
     defaultLocale: 'en',
   },
   sassOptions: {
@@ -37,6 +37,7 @@ module.exports = {
   webpack: (config) => {
     config.experiments = config.experiments || {}
     config.experiments.topLevelAwait = true
+    config.resolve.fallback = { fs: false }
     return config
   },
 
@@ -48,4 +49,20 @@ module.exports = {
   // This causes problems with serving stale content, since upon revalidating
   // the CDN marks the cached content as "fresh".
   generateEtags: false,
+
+  experimental: {
+    // The output of our getServerSideProps() return large chunks of
+    // data because it contains our rendered Markdown.
+    // The default, for a "Large Page Data" warning is 128KB
+    // but many of our pages are much larger.
+    // The warning is: https://nextjs.org/docs/messages/large-page-data
+    largePageDataBytes: 1024 * 1024, // 1 MB
+
+    // This makes it so that going Back will scroll to the previous position
+    scrollRestoration: true,
+  },
+
+  compiler: {
+    styledComponents: true,
+  },
 }
