@@ -16,8 +16,10 @@ export default [
   'Accounts',
   'Action development',
   'Actions',
+  'Actions Runner Controller',
   'Administrator',
   'Advanced Security',
+  'AI',
   'Alerts',
   'Amazon ECS',
   'Analytics',
@@ -30,6 +32,7 @@ export default [
   'Azure Kubernetes Service',
   'Azure Pipelines',
   'Azure Static Web Apps',
+  'Azure Virtual Network',
   'Backups',
   'Billing',
   'C/C++',
@@ -53,6 +56,7 @@ export default [
   'Dependencies',
   'Dependency graph',
   'Dependency review',
+  'Deployment',
   'Desktop',
   'Developer',
   'Device verification',
@@ -86,6 +90,7 @@ export default [
   'Java',
   'JavaScript',
   'Jenkins',
+  'Kotlin',
   'Legal',
   'LFS',
   'Licensing',
@@ -100,7 +105,7 @@ export default [
   'Node',
   'Nonprofits',
   'Notifications',
-  'OAuth Apps',
+  'OAuth apps',
   'Open Source',
   'Organizations',
   'Packages',
@@ -142,7 +147,9 @@ export default [
   'Storage',
   'Support',
   'Swift',
+  'Team',
   'Teams',
+  'Transcripts',
   'Travis CI',
   'Trials',
   'Troubleshooting',
